Déplacement de checkMemory dans son propre fichier
